+++ NEWS.md
@@ -1,3 +1,7 @@
+# mrf2d 0.4
+  * Added `mrfout` and `hmrfout` classes for model fitting functions output. A `summary()` and `plot()` methods are also available.
+  * Simplified structure of documentation files.
+
 # mrf2d 0.3
 
   * Co-ocurrence histogram computing is now exported in the `cohist()` function.